feat(gh): add lean-thoughts-ci collaborator
This will be used to detach my account from CI.
