+++ internal/gh/repository.go
@@ -1,0 +1,79 @@
+package gh
+
+import (
+	"github.com/pulumi/pulumi-github/sdk/v5/go/github"
+	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
+)
+
+func repository(ctx *pulumi.Context, repo *Repository) (*github.Repository, error) {
+	t, err := template(repo)
+	if err != nil {
+		return nil, err
+	}
+
+	if err := repo.Checks.Valid(); err != nil {
+		return nil, err
+	}
+
+	args := &github.RepositoryArgs{
+		AllowMergeCommit:    pulumi.Bool(false),
+		AllowRebaseMerge:    pulumi.Bool(false),
+		AllowUpdateBranch:   pulumi.Bool(true),
+		AutoInit:            pulumi.Bool(true),
+		DeleteBranchOnMerge: pulumi.Bool(true),
+		Description:         pulumi.String(repo.Description),
+		HasDownloads:        pulumi.Bool(true),
+		HasIssues:           pulumi.Bool(true),
+		HasProjects:         pulumi.Bool(true),
+		HasWiki:             pulumi.Bool(true),
+		HomepageUrl:         pulumi.String(repo.HomepageURL),
+		IsTemplate:          pulumi.Bool(repo.IsTemplate),
+		Name:                pulumi.String(repo.Name),
+		Archived:            pulumi.Bool(repo.Archived),
+		Pages:               pages(repo),
+		SecurityAndAnalysis: &github.RepositorySecurityAndAnalysisArgs{
+			SecretScanning: &github.RepositorySecurityAndAnalysisSecretScanningArgs{
+				Status: pulumi.String("enabled"),
+			},
+			SecretScanningPushProtection: &github.RepositorySecurityAndAnalysisSecretScanningPushProtectionArgs{
+				Status: pulumi.String("enabled"),
+			},
+		},
+		SquashMergeCommitTitle: pulumi.String("PR_TITLE"),
+		Template:               t,
+		Topics:                 pulumi.ToStringArray(repo.Topics),
+		Visibility:             pulumi.String(repo.Visibility),
+		VulnerabilityAlerts:    pulumi.Bool(true),
+	}
+
+	return github.NewRepository(ctx, repo.Name, args)
+}
+
+func template(repo *Repository) (*github.RepositoryTemplateArgs, error) {
+	if repo.Template == nil {
+		return nil, nil
+	}
+
+	if err := repo.Template.Valid(); err != nil {
+		return nil, err
+	}
+
+	args := &github.RepositoryTemplateArgs{
+		Owner:      pulumi.String(repo.Template.Owner),
+		Repository: pulumi.String(repo.Template.Repository),
+	}
+
+	return args, nil
+}
+
+func pages(repo *Repository) *github.RepositoryPagesArgs {
+	if !repo.EnablePages {
+		return nil
+	}
+
+	return &github.RepositoryPagesArgs{
+		Source: &github.RepositoryPagesSourceArgs{
+			Branch: pulumi.String(master),
+		},
+	}
+}
